+++ src/compiler/llvm_codegen_expr.c
@@ -7062,6 +7062,7 @@ void llvm_emit_expr(GenContext *c, BEValue *value, Expr *expr)
 		case EXPR_EMBED:
 		case EXPR_MACRO_BODY:
 		case EXPR_OTHER_CONTEXT:
+		case EXPR_DESIGNATOR:
 			UNREACHABLE
 		case EXPR_DEFAULT_ARG:
 			llvm_emit_default_arg(c, value, expr);