+++ src/compiler/sema_expr.c
@@ -588,6 +588,7 @@ static bool sema_binary_is_expr_lvalue(SemaContext *context, Expr *top_expr, Exp
 		case EXPR_SLICE_ASSIGN:
 		case EXPR_SLICE_COPY:
 		case EXPR_STRINGIFY:
+		case EXPR_TAGOF:
 		case EXPR_TERNARY:
 		case EXPR_TRY_UNWRAP:
 		case EXPR_TRY_UNWRAP_CHAIN:
@@ -618,6 +619,7 @@ static bool expr_may_ref(Expr *expr)
 		case EXPR_CT_IDENT:
 		case EXPR_EMBED:
 		case EXPR_DEFAULT_ARG:
+		case EXPR_TAGOF:
 			return false;
 		case EXPR_OTHER_CONTEXT:
 			return expr_may_ref(expr->expr_other_context.inner);
@@ -2450,6 +2452,7 @@ bool sema_expr_analyse_general_call(SemaContext *context, Expr *expr, Decl *decl
 		                                  type_flatten(exprptr(expr->call_expr.function)->type), optional,
 		                                  no_match_ref);
 	}
+	if (!sema_analyse_decl(context, decl)) return false;
 	switch (decl->decl_kind)
 	{
 		case DECL_MACRO:
@@ -2476,11 +2479,52 @@ bool sema_expr_analyse_general_call(SemaContext *context, Expr *expr, Decl *decl
 	}
 }
 
+static inline bool sema_expr_analyse_tagof(SemaContext *context, Expr *expr)
+{
+	expr->call_expr.arguments = sema_expand_vasplat_exprs(context, expr->call_expr.arguments);
+	Expr **args = expr->call_expr.arguments;
+	unsigned arg_count = vec_size(args);
+	Expr *tag = exprptr(expr->call_expr.function);
+	Decl *decl = tag->tag_of_expr.type;
+	bool is_has = tag->tag_of_expr.property == TYPE_PROPERTY_HAS_TAGOF;
+	const char *name = is_has ? "has_tagof" : "tagof";
+	if (arg_count != 1) RETURN_SEMA_ERROR(expr, "Expected a single string argument to '%s'.", name);
+	Expr *key = args[0];
+	if (!sema_analyse_expr(context, key)) return false;
+	if (!sema_cast_const(key) || !expr_is_const_string(key))
+	{
+		RETURN_SEMA_ERROR(key, "The tag name should be a string constant.");
+	}
+	assert(decl->resolved_attributes);
+	ResolvedAttrData *attrs = decl->attrs_resolved;
+	if (!attrs || !attrs->tags) goto NOT_FOUND;
+	const char *tagname = key->const_expr.bytes.ptr;
+	Expr *value = NULL;
+	FOREACH(Attr *, attr, attrs->tags)
+	{
+		if (str_eq(attr->exprs[0]->const_expr.bytes.ptr, tagname)) value = attr->exprs[1];
+	}
+	if (!value) goto NOT_FOUND;
+	if (is_has)
+	{
+		expr_rewrite_const_bool(expr, type_bool, true);
+		return true;
+	}
+	expr_replace(expr, expr_copy(value));
+	return true;
+NOT_FOUND:
+	if (is_has)
+	{
+		expr_rewrite_const_bool(expr, type_bool, false);
+		return true;
+	}
+	RETURN_SEMA_ERROR(expr, "The tag '%s' is not defined, always check with '.has_tagof'.");
+}
+
 static inline bool sema_expr_analyse_call(SemaContext *context, Expr *expr, bool *no_match_ref)
 {
 	if (no_match_ref) *no_match_ref = true;
 	Expr *func_expr = exprptr(expr->call_expr.function);
-
 	if (!sema_analyse_expr_lvalue_fold_const(context, func_expr)) return false;
 	if (func_expr->expr_kind == EXPR_MACRO_BODY_EXPANSION)
 	{
@@ -2491,6 +2535,8 @@ static inline bool sema_expr_analyse_call(SemaContext *context, Expr *expr, bool
 	Expr *struct_var = NULL;
 	switch (func_expr->expr_kind)
 	{
+		case EXPR_TAGOF:
+			return sema_expr_analyse_tagof(context, expr);
 		case EXPR_BUILTIN:
 			return sema_expr_analyse_builtin_call(context, expr);
 		case EXPR_IDENTIFIER:
@@ -3391,6 +3437,11 @@ static inline bool sema_expr_analyse_member_access(SemaContext *context, Expr *e
 	TypeProperty type_property = type_property_by_name(name);
 	switch (type_property)
 	{
+		case TYPE_PROPERTY_TAGOF:
+		case TYPE_PROPERTY_HAS_TAGOF:
+			expr->expr_kind = EXPR_TAGOF;
+			expr->tag_of_expr = (ExprTagOf) { .type = decl, .property = type_property };
+			return true;
 		case TYPE_PROPERTY_NONE:
 			break;
 		case TYPE_PROPERTY_QNAMEOF:
@@ -3869,6 +3920,8 @@ static bool sema_expr_rewrite_to_typeid_property(SemaContext *context, Expr *exp
 		case TYPE_PROPERTY_MEMBERSOF:
 		case TYPE_PROPERTY_METHODSOF:
 		case TYPE_PROPERTY_EXTNAMEOF:
+		case TYPE_PROPERTY_TAGOF:
+		case TYPE_PROPERTY_HAS_TAGOF:
 		case TYPE_PROPERTY_NAMEOF:
 		case TYPE_PROPERTY_QNAMEOF:
 		case TYPE_PROPERTY_ASSOCIATED:
@@ -4117,6 +4170,8 @@ static bool sema_type_property_is_valid_for_type(Type *original_type, TypeProper
 		case TYPE_PROPERTY_PARAMS:
 		case TYPE_PROPERTY_RETURNS:
 			return type_is_func_ptr(type);
+		case TYPE_PROPERTY_TAGOF:
+		case TYPE_PROPERTY_HAS_TAGOF:
 		case TYPE_PROPERTY_EXTNAMEOF:
 			return !type_is_builtin(type->type_kind);
 	}
@@ -4221,6 +4276,11 @@ static bool sema_expr_rewrite_to_type_property(SemaContext *context, Expr *expr,
 			assert(!type_is_builtin(type->type_kind));
 			if (!sema_resolve_type_decl(context, type)) return false;
 			sema_expr_rewrite_to_type_nameof(expr, type, TOKEN_CT_EXTNAMEOF);
+			return true;
+		case TYPE_PROPERTY_TAGOF:
+		case TYPE_PROPERTY_HAS_TAGOF:
+			expr->expr_kind = EXPR_TAGOF;
+			expr->tag_of_expr = (ExprTagOf) { .type = type->decl, .property = property };
 			return true;
 		case TYPE_PROPERTY_NONE:
 			return false;
@@ -8876,6 +8936,7 @@ static inline bool sema_expr_analyse_ct_defined(SemaContext *context, Expr *expr
 			case EXPR_POST_UNARY:
 			case EXPR_TYPEID:
 			case EXPR_TYPEID_INFO:
+			case EXPR_TAGOF:
 				if (!sema_analyse_expr(active_context, main_expr)) return false;
 				break;
 		}
@@ -8901,7 +8962,7 @@ static inline bool sema_expr_analyse_ct_arg(SemaContext *context, Type *infer_ty
 			return true;
 		case TOKEN_CT_VAARG:
 		{
-			unsigned index;
+			unsigned index = 0;
 			// A normal argument, this means we only evaluate it once.
 			ASSIGN_EXPR_OR_RET(Expr *arg_expr, sema_expr_analyse_ct_arg_index(context, exprptr(expr->ct_arg_expr.arg),
 			                                                                  &index, true), false);
@@ -8913,7 +8974,7 @@ static inline bool sema_expr_analyse_ct_arg(SemaContext *context, Type *infer_ty
 			FOREACH(Decl *, val, context->macro_params)
 			{
 				if (!val) continue;
-				if (val->va_index == index && val->var.kind == VARDECL_PARAM)
+				if (val->var.va_index == index && val->var.kind == VARDECL_PARAM)
 				{
 					decl = val;
 					break;
@@ -8929,7 +8990,7 @@ static inline bool sema_expr_analyse_ct_arg(SemaContext *context, Type *infer_ty
 				}
 				decl = decl_new_generated_var(arg_expr->type, VARDECL_PARAM, arg_expr->span);
 				decl->var.init_expr = arg_expr;
-				decl->va_index = (uint16_t)index;
+				decl->var.va_index = (uint16_t)index;
 				vec_add(context->macro_params, decl);
 			}
 			// Replace with the identifier.
@@ -8976,7 +9037,7 @@ static inline bool sema_expr_analyse_ct_arg(SemaContext *context, Type *infer_ty
 			FOREACH(Decl *, val, context->macro_params)
 			{
 				if (!val) continue;
-				if (val->var.kind == VARDECL_PARAM_REF && val->va_index == index)
+				if (val->var.kind == VARDECL_PARAM_REF && val->var.va_index == index)
 				{
 					decl = val;
 					break;
@@ -8990,7 +9051,7 @@ static inline bool sema_expr_analyse_ct_arg(SemaContext *context, Type *infer_ty
 				expr_insert_addr(arg_expr);
 				decl = decl_new_generated_var(arg_expr->type, VARDECL_PARAM_REF, arg_expr->span);
 				decl->var.init_expr = arg_expr;
-				decl->va_index = (uint16_t)index;
+				decl->var.va_index = (uint16_t)index;
 				vec_add(context->macro_params, decl);
 			}
 			// Replace with the identifier.
@@ -9493,6 +9554,7 @@ static inline bool sema_expr_analyse_builtin(SemaContext *context, Expr *expr, b
 	return true;
 }
 
+
 static inline bool sema_expr_analyse_compound_literal(SemaContext *context, Expr *expr)
 {
 	TypeInfo *type_info = expr->expr_compound_literal.type_info;
@@ -9532,6 +9594,8 @@ static inline bool sema_analyse_expr_dispatch(SemaContext *context, Expr *expr)
 		case EXPR_MACRO_BODY:
 		case EXPR_DEFAULT_ARG:
 			UNREACHABLE
+		case EXPR_TAGOF:
+			RETURN_SEMA_ERROR(expr, "Expected '()' after this.");
 		case EXPR_OTHER_CONTEXT:
 			context = expr->expr_other_context.context;
 			expr_replace(expr, expr->expr_other_context.inner);
@@ -9837,19 +9901,18 @@ static inline bool sema_cast_rvalue(SemaContext *context, Expr *expr)
 				return false;
 			}
 			break;
+		case EXPR_TAGOF:
+			RETURN_SEMA_ERROR(expr, "A tag name must be given.");
 		case EXPR_BUILTIN:
-			SEMA_ERROR(expr, "A builtin must be followed by ().");
-			return false;
+			RETURN_SEMA_ERROR(expr, "A builtin must be followed by ().");
 		case EXPR_ACCESS:
 			if (expr->access_expr.ref->decl_kind == DECL_FUNC)
 			{
-				SEMA_ERROR(expr, "A function name must be followed by '(' or preceded by '&'.");
-				return false;
+				RETURN_SEMA_ERROR(expr, "A function name must be followed by '(' or preceded by '&'.");
 			}
 			if (expr->access_expr.ref->decl_kind == DECL_MACRO)
 			{
-				SEMA_ERROR(expr, "A macro name must be followed by '('.");
-				return false;
+				RETURN_SEMA_ERROR(expr, "A macro name must be followed by '('.");
 			}
 			// We may have kept FOO.x.y as a reference, fold it now if y is not an aggregate.
 			sema_expr_flatten_const_ident(expr->access_expr.parent);