+++ src/compiler/sema_stmts.c
@@ -631,6 +631,7 @@ static inline bool sema_expr_valid_try_expression(Expr *expr)
 		case EXPR_POST_UNARY:
 		case EXPR_TERNARY:
 		case EXPR_LAST_FAULT:
+		case EXPR_TAGOF:
 			return false;
 		case EXPR_BITACCESS:
 		case EXPR_BUILTIN: