+++ releasenotes.md
@@ -24,7 +24,8 @@
 - Add `@const` attribute for macros, for better error messages with constant macros.
 - Add `wincrt` setting to libraries.
 - Add `+++` `&&&` `|||` as replacement for `$concat`, `$and` and `$or`.
-- Add `methodsof` to type info for struct, union and bitstruct
+- Add `methodsof` to type info for struct, union and bitstruct.
+- Added `@tag` `tagof` and `has_tagof` to user defined types and members.
 
 ### Fixes
 
@@ -50,6 +51,7 @@
 - Improved output when pointer is out of range.
 - Better error when casting to a distinct fails.
 - With single module, name the .o file after what `-o` provides. #1306
+- Bitstruct members can now have attributes.
 
 ### Stdlib changes